# api/openapi.yaml
openapi: 3.0.3
info:
  title: Inditex ZARA Boost workshop
  description: |-
    Inditex ZARA Boost workshop
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /ranking:
    get:
      tags:
        - ranking
      summary: Product Order Ranking
      operationId: ranking
      parameters:
        - name: type
          in: query
          description: Ranking Type
          required: true
          explode: false
          schema:
            $ref: '#/components/schemas/RankingType'
        - name: fromDate
          in: query
          description: Date from which to start product sales analysis.
          required: true
          schema:
            type: string
            format: date
            example: "2023-01-01"
        - name: toDate
          in: query
          description: Limit date for the product sales analysis.
          required: true
          schema:
            type: string
            format: date
            example: "2023-01-15"        
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ranking'
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Product 1
        price:
          type: number
          format: double
          example: 39.95
        category:
          type: string
          example: Pants
        imageUrl:
          type: string
          example: /assets/pants_a.jpg
...
    RankingType:
      type: string
      description: Ranking Type
      example: orders
      enum:
        - orders
        - units
        - amount
    Ranking:
      type: object
      properties:
        ranking-type:
          $ref: '#/components/schemas/RankingType'
        productsRank:
          type: array
          items:
            $ref: '#/components/schemas/ProductRank'
    ProductRank:
      allOf:
        - $ref: '#/components/schemas/Product'
      type: object
      properties:
        rank:
          type: integer
        score:
          type: number
...